requirements/base.txt
Django>=1.9
django-environ>=0.4.0
django-braces>=1.8.1
django-crispy-forms>=1.5.2
django-admin-bootstrapped>=2.5.6
djangorestframework>=3.3.2
django-filter>=0.13.0
django-cors-headers>=1.1.0
django-allauth>=0.24.1
django-storages_redux>=1.3
django-recaptcha>=1.0.4
pytz>=2016.1
redis>=2.10.3
hiredis>=0.2.0
django-redis-cache>=1.6.3
boto3>=1.3.0
boto>=2.39.0



